pyproject.toml
[project]
name = "ncs-experiment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "keras>=3.11.3",
    "keras-cv>=0.9.0",
    "keras-hub>=0.19.0",
    "onnx>=1.19.0",
    "opencv-python>=4.12.0.88",
    "openvino>=2025.3.0",
    "pillow>=11.3.0",
    "tensorflow>=2.20.0",
    "tf2onnx>=1.8.4",
    "ultralytics>=8.3.203",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
